decode-sse: Fix PSHUFD/SHUFPS/SHUFPD
When the instructions PSHUFD, SHUFPS, or SHUFPD are executed with the
same source (rareg) and destination register (rdreg), the lower part is
already overwritten at the time the permutation of the higher part is
performed, leading to incorrect results.

This is fixed by working on temporary registers and adding extra move
uOps afterwards.